DD-2009 Consistency checks dd-data-vault (#13)
Refactored a number of interface methods that took a path to a layer, which was in fact always a layer ID. These methods now take a layerId of type long.
Added methods to retrieve the list of layer IDs from the database and from the layer as it exists in storage.
Added nl.knaw.dans.layerstore.LayeredItemStore#checkSameLayerIdsFoundOnStorageAsInDatabase which throws an IllegalStateException if the DB is inconsitent with storage.